requirements.txt
Django==5.1.1
django-extensions==3.2.3
djangorestframework==3.15.2
pillow==10.4.0